Fixed transaction file and testing file is also working
