+++ ae1Environment/card-checker/pom.xml
@@ -1,29 +1,16 @@
 <?xml version="1.0" encoding="UTF-8"?>
 <project xmlns="http://maven.apache.org/POM/4.0.0" xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance" xsi:schemaLocation="http://maven.apache.org/POM/4.0.0 http://maven.apache.org/xsd/maven-4.0.0.xsd">
-
+    <modelVersion>4.0.0</modelVersion>
+    
     <parent>
         <groupId>org.solent.ood</groupId>
         <artifactId>creditcardchecker</artifactId>
         <version>0.1-SNAPSHOT</version>
-    </parent> 
-
-    <modelVersion>4.0.0</modelVersion>
+    </parent>
+    
+   
     <artifactId>card-checker</artifactId>
     <packaging>jar</packaging>
-
-   <build>
-        <plugins>
-            <plugin>
-                <groupId>org.apache.maven.plugins</groupId>
-                <artifactId>maven-compiler-plugin</artifactId>
-                <version>3.1</version>
-                <configuration>
-                    <source>1.7</source>
-                    <target>1.7</target>
-                </configuration>
-            </plugin>
-</plugins>
-</build>
     
     <dependencies>
         <dependency>
