--- ae1Environment/card-checker/src/main/java/solent/ac/uk/ood/examples/cardcheck/CalculateLunnDigit.java
@@ -1,62 +1,0 @@
-/* Modifications Copyright 2018 Craig Gallen
-  *********************************************
-  * Original https://www.admfactory.com/luhn-algorithm-implementation-in-java/
-  * Licensed under the Apache License, Version 2.0 (the "License");
-  * you may not use this file except in compliance with the License.
-  * You may obtain a copy of the License at
-
-  * http://www.apache.org/licenses/LICENSE-2.0
-
-  * Unless required by applicable law or agreed to in writing, software
-  * distributed under the License is distributed on an "AS IS" BASIS,
-  * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-  * See the License for the specific language governing permissions and
-  * limitations under the License.
- */
-package solent.ac.uk.ood.examples.cardcheck;
-
-/**
- *
- * @author cgallen
- */
-public class CalculateLunnDigit {
-
-    /**
-     * Calculates the last digits for the card number received as parameter
-     *
-     * @param card number
-     * @return the check digit
-     */
-    public static String calculateCheckDigit(String card) {
-        if (card == null) {
-            return null;
-        }
-        String digit;
-        /* convert to array of int for simplicity */
-        int[] digits = new int[card.length()];
-        for (int i = 0; i < card.length(); i++) {
-            digits[i] = Character.getNumericValue(card.charAt(i));
-        }
-
-        /* double every other starting from right - jumping from 2 in 2 */
-        for (int i = digits.length - 1; i >= 0; i -= 2) {
-            digits[i] += digits[i];
-
-            /* taking the sum of digits grater than 10 - simple trick by substract 9 */
-            if (digits[i] >= 10) {
-                digits[i] = digits[i] - 9;
-            }
-        }
-        int sum = 0;
-        for (int i = 0; i < digits.length; i++) {
-            sum += digits[i];
-        }
-        /* multiply by 9 step */
-        sum = sum * 9;
-
-        /* convert to string to be easier to take the last digit */
-        digit = sum + "";
-        return digit.substring(digit.length() - 1);
-    }
-
-}
