+++ ae1Environment/client/pom.xml
@@ -2,16 +2,15 @@
 <project xmlns="http://maven.apache.org/POM/4.0.0" xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance" xsi:schemaLocation="http://maven.apache.org/POM/4.0.0 http://maven.apache.org/xsd/maven-4.0.0.xsd">
     <modelVersion>4.0.0</modelVersion>
     
-    <groupId>solent.ac.uk.ood.examples</groupId>
-    <artifactId>bank-client</artifactId>
-    <version>1.0-SNAPSHOT</version>
-    <packaging>jar</packaging>
+    <parent>
+        <groupId>org.solent.ood</groupId>
+        <artifactId>creditcardchecker</artifactId>
+        <version>0.1-SNAPSHOT</version>
+    </parent>
     
-    <properties>
-        <project.build.sourceEncoding>UTF-8</project.build.sourceEncoding>
-        <maven.compiler.source>1.8</maven.compiler.source>
-        <maven.compiler.target>1.8</maven.compiler.target>
-    </properties>
+   
+    <artifactId>bank-client</artifactId>
+    <packaging>jar</packaging>
     
     <build>
         <plugins>
