--- ae1Environment/client/src/main/java/org/solent/com504/oodd/bank/model/client/BankRestClient.java
@@ -1,11 +1,0 @@
-package org.solent.com504.oodd.bank.model.client;
-
-import org.solent.com504.oodd.bank.model.dto.CreditCard;
-import org.solent.com504.oodd.bank.model.dto.TransactionReplyMessage;
-
-public interface BankRestClient {
-
-    public  TransactionReplyMessage transferMoney(CreditCard fromCard, CreditCard toCard, Double amount);
-
-    public  TransactionReplyMessage transferMoney(CreditCard fromCard, CreditCard toCard, Double amount, String userName, String password);
-}
