--- ae1Environment/client/src/main/java/org/solent/com504/oodd/bank/model/dto/BankTransactionStatus.java
@@ -1,6 +1,0 @@
-package org.solent.com504.oodd.bank.model.dto;
-
-public enum BankTransactionStatus {
-
-    SUCCESS, FAIL
-}
