--- ae1Environment/client/src/main/java/org/solent/com504/oodd/password/PasswordUtils.java
@@ -1,24 +1,0 @@
-/*
- * To change this license header, choose License Headers in Project Properties.
- * To change this template file, choose Tools | Templates
- * and open the template in the editor.
- */
-package org.solent.com504.oodd.password;
-
-import org.mindrot.jbcrypt.BCrypt;
-
-/**
- *
- * @author gallenc
- */
-public class PasswordUtils {
-    
-    public static String hashPassword(String password){
-        return BCrypt.hashpw(password, BCrypt.gensalt(12));
-    }
-    
-    public static boolean checkPassword(String password, String hashed){
-        return BCrypt.checkpw(password, hashed);
-    }
-    
-}
