+++ ae1Environment/client/src/main/java/org/solent/odd/bank/model/client/BankRestClient.java
@@ -1,0 +1,11 @@
+package org.solent.ood.bank.model.client;
+
+import org.solent.ood.bank.model.dto.CreditCard;
+import org.solent.ood.bank.model.dto.TransactionReplyMessage;
+
+public interface BankRestClient {
+
+    public  TransactionReplyMessage transferMoney(CreditCard fromCard, CreditCard toCard, Double amount);
+
+    public  TransactionReplyMessage transferMoney(CreditCard fromCard, CreditCard toCard, Double amount, String userName, String password);
+}
