+++ ae1Environment/client/src/main/java/org/solent/odd/bank/model/dto/BankTransactionStatus.java
@@ -1,0 +1,6 @@
+package org.solent.ood.bank.model.dto;
+
+public enum BankTransactionStatus {
+
+    SUCCESS, FAIL
+}
