+++ ae1Environment/client/src/main/java/org/solent/odd/bank/model/dto/CreditCard.java
@@ -1,0 +1,66 @@
+package org.solent.ood.bank.model.dto;
+
+import javax.persistence.Embeddable;
+
+@Embeddable
+public class CreditCard {
+
+    private String name="";
+
+    private String endDate="";
+
+    private String cardnumber="";
+
+    private String cvv="111";
+
+    private String issueNumber="01";
+
+    public String getName() {
+        return name;
+    }
+
+    public void setName(String name) {
+        this.name = name;
+    }
+
+    public String getEndDate() {
+        return endDate;
+    }
+
+    public void setEndDate(String endDate) {
+        this.endDate = endDate;
+    }
+
+    public String getCardnumber() {
+        return cardnumber;
+    }
+
+    public void setCardnumber(String cardnumber) {
+        this.cardnumber = cardnumber;
+    }
+
+    public String getCvv() {
+        return cvv;
+    }
+
+    public void setCvv(String cvv) {
+        this.cvv = cvv;
+    }
+
+    public String getIssueNumber() {
+        return issueNumber;
+    }
+
+    public void setIssueNumber(String issueNumber) {
+        this.issueNumber = issueNumber;
+    }
+
+    @Override
+    public String toString() {
+        return "CreditCard{" + "name=" + name + ", endDate=" + endDate + ", cardnumber=" + cardnumber + ", cvv=NOT PRINTED" +  ", issueNumber=" + issueNumber + '}';
+    }
+    
+    
+    
+    
+}
