+++ ae1Environment/client/src/main/java/org/solent/odd/bank/model/dto/TransactionReplyMessage.java
@@ -1,0 +1,97 @@
+package org.solent.ood.bank.model.dto;
+
+import java.util.Date;
+
+
+public class TransactionReplyMessage {
+
+    private Integer code;
+
+    private String message;
+
+     private BankTransactionStatus status;
+
+    private String fromCardNo;
+
+    private String toCardNo;
+
+    private String transactionId;
+
+    private Date transactionDate;
+    
+    private Double amount;
+
+    public Integer getCode() {
+        return code;
+    }
+
+    public void setCode(Integer code) {
+        this.code = code;
+    }
+
+    public String getMessage() {
+        return message;
+    }
+
+    public void setMessage(String message) {
+        this.message = message;
+    }
+
+    public BankTransactionStatus getStatus() {
+        return status;
+    }
+
+    public void setStatus(BankTransactionStatus status) {
+        this.status = status;
+    }
+
+    public String getFromCardNo() {
+        return fromCardNo;
+    }
+
+    public void setFromCardNo(String fromCardNo) {
+        this.fromCardNo = fromCardNo;
+    }
+
+    public String getToCardNo() {
+        return toCardNo;
+    }
+
+    public void setToCardNo(String toCardNo) {
+        this.toCardNo = toCardNo;
+    }
+
+    public String getTransactionId() {
+        return transactionId;
+    }
+
+    public void setTransactionId(String transactionId) {
+        this.transactionId = transactionId;
+    }
+
+    public Date getTransactionDate() {
+        return transactionDate;
+    }
+
+    public void setTransactionDate(Date transactionDate) {
+        this.transactionDate = transactionDate;
+    }
+
+    public Double getAmount() {
+        return amount;
+    }
+
+    public void setAmount(Double amount) {
+        this.amount = amount;
+    }
+
+    @Override
+    public String toString() {
+        return "TransactionReplyMessage{" + "code=" + code + ", message=" + message + ", status=" + status + ", fromCardNo=" + fromCardNo + ", toCardNo=" + toCardNo + ", transactionId=" + transactionId + ", transactionDate=" + transactionDate + ", amount=" + amount + '}';
+    }
+
+
+    
+    
+    
+}
