--- ae1Environment/web-ui/README.md
@@ -1,1 +1,0 @@
-This is where the UI interface of the project will be in
